#102 add TaskDetails to TaskList
